requirements.txt
instagrapi==1.16.2
python-dotenv==0.19.2
